Remove height setting on main element
The h-screen class was causing the height of the <main> element to be
fixed at a given size which did not track with the content of the box.

This change removes this layout bug but introduces the undesired
behaviour that the footer is at differing positions on the screen
depending on the content of the page.

Need to set up a 'sticky' footer using the tailwind css classes.